Import freesurfer 5.3+cvs20130513

+++ freeview/MyCmdLineParser.cpp
@@ -6,9 +6,9 @@
 /*
  * Original Author: Ruopeng Wang
  * CVS Revision Info:
- *    $Author: rpwang $
- *    $Date: 2011/05/13 15:04:32 $
- *    $Revision: 1.17.2.2 $
+ *    $Author: zkaufman $
+ *    $Date: 2013/05/03 17:52:34 $
+ *    $Revision: 1.17.2.8 $
  *
  * Copyright © 2011 The General Hospital Corporation (Boston, MA) "MGH"
  *
@@ -137,7 +137,7 @@ bool MyCmdLineParser::Found( const QString flag, QStringList* args, int nIndex )
   if (ret)
   {
     args->clear();
-    for ( int i = 0; i < sa.size(); i++ )
+    for ( size_t i = 0; i < sa.size(); i++ )
     {
       args->push_back( sa[i].c_str() );
     }
@@ -148,7 +148,7 @@ bool MyCmdLineParser::Found( const QString flag, QStringList* args, int nIndex )
 bool MyCmdLineParser::Found( const char* chFlag, CmdLineEntry* e, int nIndex )
 {
   int n = 0;
-  for ( int i = 0; i < m_cmdLineEntries.size(); i++ )
+  for ( size_t i = 0; i < m_cmdLineEntries.size(); i++ )
   {
     if ( strcmp( m_cmdLineEntries[i].shortName, chFlag ) == 0 ||
          strcmp( m_cmdLineEntries[i].longName, chFlag ) == 0 )
